# perf: reduce if statement for `NaN` (micro-optimization)