--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -25,11 +25,11 @@ module.exports = class Serializer {
 
   asInteger (i) {
     if (typeof i === 'number') {
-      if (i === Infinity || i === -Infinity || Number.isNaN(i)) {
-        throw new Error(`The value "${i}" cannot be converted to an integer.`)
-      }
       if (Number.isInteger(i)) {
         return '' + i
+      }
+      if (i === Infinity || i === -Infinity || Number.isNaN(i)) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
       }
       return this.parseInteger(i)
     } else if (i === null) {