--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -28,7 +28,8 @@ module.exports = class Serializer {
       if (Number.isInteger(i)) {
         return '' + i
       }
-      if (i === Infinity || i === -Infinity || Number.isNaN(i)) {
+      // check if number is Infinity or NaN
+      if (i === Infinity || i === -Infinity || i !== i) {
         throw new Error(`The value "${i}" cannot be converted to an integer.`)
       }
       return this.parseInteger(i)