--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -29,7 +29,7 @@ module.exports = class Serializer {
         return '' + i
       }
       // check if number is Infinity or NaN
-      // eslint-disable-next-line
+      // eslint-disable-next-line no-self-compare
       if (i === Infinity || i === -Infinity || i !== i) {
         throw new Error(`The value "${i}" cannot be converted to an integer.`)
       }