--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -51,7 +51,9 @@ module.exports = class Serializer {
 
   asNumber (i) {
     const num = Number(i)
-    if (Number.isNaN(num)) {
+    // check if number is NaN
+    // eslint-disable-next-line no-self-compare
+    if (num !== num) {
       throw new Error(`The value "${i}" cannot be converted to a number.`)
     } else if (!Number.isFinite(num)) {
       return 'null'