@cesco69 2024-03-13
@ivan-tymoshenko the difference is very small

before (master)
```
short string............................................. x 21,651,330 ops/sec ±1.82% (185 runs sampled)
unsafe short string...................................... x 890,279,050 ops/sec ±1.42% (181 runs sampled)
short string with double quote........................... x 12,229,862 ops/sec ±2.59% (190 runs sampled)
long string without double quotes........................ x 15,138 ops/sec ±0.37% (190 runs sampled)
unsafe long string without double quotes................. x 1,022,578,589 ops/sec ±0.39% (186 runs sampled)
long string.............................................. x 15,864 ops/sec ±0.47% (189 runs sampled)
unsafe long string....................................... x 1,054,094,454 ops/sec ±0.47% (190 runs sampled)
number................................................... x 1,060,238,499 ops/sec ±0.43% (190 runs sampled)
integer.................................................. x 219,863,056 ops/sec ±0.41% (189 runs sampled)
formatted date-time...................................... x 1,708,781 ops/sec ±0.39% (191 runs sampled)
formatted date........................................... x 1,140,619 ops/sec ±0.41% (191 runs sampled)
formatted time........................................... x 1,139,486 ops/sec ±0.38% (191 runs sampled)
short array of numbers................................... x 74,560 ops/sec ±0.47% (189 runs sampled)
short array of integers.................................. x 62,719 ops/sec ±0.45% (189 runs sampled)
short array of short strings............................. x 20,168 ops/sec ±0.41% (191 runs sampled)
short array of long strings.............................. x 19,610 ops/sec ±0.64% (191 runs sampled)
short array of objects with properties of different types x 9,593 ops/sec ±0.48% (190 runs sampled)
object with number property.............................. x 1,039,195,747 ops/sec ±0.38% (190 runs sampled)
object with integer property............................. x 221,921,570 ops/sec ±0.42% (187 runs sampled)
object with short string property........................ x 22,879,162 ops/sec ±0.62% (191 runs sampled)
object with long string property......................... x 16,001 ops/sec ±0.34% (191 runs sampled)
object with properties of different types................ x 1,928,181 ops/sec ±0.84% (185 runs sampled)
simple object............................................ x 9,957,657 ops/sec ±0.41% (192 runs sampled)
simple object with required fields....................... x 9,954,660 ops/sec ±0.41% (191 runs sampled)
object with const string property........................ x 1,065,629,415 ops/sec ±0.40% (191 runs sampled)
object with const number property........................ x 1,050,253,605 ops/sec ±0.45% (188 runs sampled)
object with const bool property.......................... x 1,065,744,689 ops/sec ±0.42% (190 runs sampled)
object with const object property........................ x 1,060,911,654 ops/sec ±0.48% (189 runs sampled)
object with const null property.......................... x 1,048,597,601 ops/sec ±0.36% (191 runs sampled)
```
after (PR)
```
short string............................................. x 23,704,625 ops/sec ±0.43% (192 runs sampled)
unsafe short string...................................... x 1,052,116,318 ops/sec ±0.45% (187 runs sampled)
short string with double quote........................... x 13,418,664 ops/sec ±0.41% (191 runs sampled)
long string without double quotes........................ x 15,433 ops/sec ±0.41% (190 runs sampled)
unsafe long string without double quotes................. x 1,046,654,174 ops/sec ±0.37% (191 runs sampled)
long string.............................................. x 16,078 ops/sec ±0.41% (191 runs sampled)
unsafe long string....................................... x 1,045,056,985 ops/sec ±0.34% (190 runs sampled)
number................................................... x 1,059,855,559 ops/sec ±0.41% (192 runs sampled)
integer.................................................. x 219,744,959 ops/sec ±0.41% (191 runs sampled)
formatted date-time...................................... x 1,702,920 ops/sec ±0.34% (192 runs sampled)
formatted date........................................... x 1,145,765 ops/sec ±0.42% (191 runs sampled)
formatted time........................................... x 1,146,381 ops/sec ±0.37% (191 runs sampled)
short array of numbers................................... x 73,906 ops/sec ±0.45% (190 runs sampled)
short array of integers.................................. x 58,721 ops/sec ±0.37% (190 runs sampled)
short array of short strings............................. x 19,936 ops/sec ±0.48% (189 runs sampled)
short array of long strings.............................. x 20,276 ops/sec ±0.45% (189 runs sampled)
short array of objects with properties of different types x 9,629 ops/sec ±0.48% (191 runs sampled)
object with number property.............................. x 1,050,962,467 ops/sec ±0.49% (189 runs sampled)
object with integer property............................. x 222,334,876 ops/sec ±0.32% (191 runs sampled)
object with short string property........................ x 22,828,654 ops/sec ±0.68% (190 runs sampled)
object with long string property......................... x 16,029 ops/sec ±0.32% (192 runs sampled)
object with properties of different types................ x 1,918,710 ops/sec ±0.89% (185 runs sampled)
simple object............................................ x 9,822,513 ops/sec ±0.46% (188 runs sampled)
simple object with required fields....................... x 9,778,720 ops/sec ±0.55% (191 runs sampled)
object with const string property........................ x 1,052,599,498 ops/sec ±0.38% (191 runs sampled)
object with const number property........................ x 1,068,877,579 ops/sec ±0.44% (192 runs sampled)
object with const bool property.......................... x 1,053,391,814 ops/sec ±0.41% (190 runs sampled)
object with const object property........................ x 1,049,257,020 ops/sec ±0.39% (189 runs sampled)
object with const null property.......................... x 1,069,585,019 ops/sec ±0.42% (190 runs sampled)
```