@cesco69 2024-03-14
> why are strings effected in the benchmark but integer and number not?

@Uzlopak I think the reason is because string serialization is CPU intensive (more complex) and instable (maybe times change on each benchmark run). Integers and Numbers are super simple block of code.

My PR can't influence strings and is a too little optimization for see big changes into "number" and "integer" benchmark time.
it is still a small optimization even if the times do not change.

If we want talk about a huge optimization, this is a good PR https://github.com/fastify/fast-json-stringify/pull/696