@Uzlopak 2024-03-14
Maybe it is better to optimize for the valid case and check first if is an integer and if not, just throw that error.