@cesco69 2024-03-14
> I dont see the performance gain? Which of these benchmarks is for you significant?

is too small optimization to see changes into the benchmark... but this is the significant part 

before (master)
```
number................................................... x 1,060,238,499 ops/sec ±0.43% (190 runs sampled)
integer.................................................. x 219,863,056 ops/sec ±0.41% (189 runs sampled)
```
after (PR)
```
number................................................... x 1,059,855,559 ops/sec ±0.41% (192 runs sampled)
integer.................................................. x 219,744,959 ops/sec ±0.41% (191 runs sampled)
```